refactor: remove to QwikCityProvider
