+++ packages/qwik/src/core/render/types.ts
@@ -36,3 +36,18 @@ export interface RenderStaticContext {
  * @public
  */
 export interface RenderContext2 {}
+
+// Polyfills for ViewTransition API
+declare global {
+  interface ViewTransition {
+    ready: Promise<void>;
+    finished: Promise<void>;
+    updateCallbackDone: Promise<void>;
+    skipTransition: () => void;
+  }
+
+  interface Document {
+    startViewTransition?: (callback: () => void | Promise<void>) => ViewTransition;
+    __q_view_transition__?: true | undefined;
+  }
+}
